\name{QCRep}
\alias{QCRep}
\docType{package}
\title{Generate a subdirectory,which is composed of QC Reports for Illuminas Infinium HumanMethylation27 BeadChip methylation assay}
\description{
QCRep is a subfunction in the main function Meth27QC}
\usage{
QCRep(Dir,controlfile,sampfile)}
\arguments{
      \item{Dir}{directory path where the control profile and sample file are included}
      \item{controlfile}{control probe file name,exported from BeadStudio/GenomeStudio}
      \item{sampfile}{sample table file name,exported from BeadStudio/GenomeStudio}

}

